pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webscrape"
version = "0.1.0"
description = "Async web scraper with httpx, BeautifulSoup, rate limiting, retry, robots.txt compliance, and rich CLI output."
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [{ name = "devaloi" }]

dependencies = [
    "httpx[http2]>=0.28",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "ruff>=0.8",
]

[project.scripts]
webscrape = "webscrape.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
